split dap command arg formatting into microcrate
