+++ Cargo.toml
@@ -90,6 +90,7 @@ members = [
     "crates/perl-dap-breakpoint",
     "crates/perl-dap-eval",
     "crates/perl-dap-platform",
+    "crates/perl-dap-command-args",
     "crates/perl-dap-stack",
     "crates/perl-dap-value",
     "crates/perl-dap-variables",
@@ -186,6 +187,7 @@ allow = [
     # Tier 5 — DAP components
     "perl-dap-breakpoint",
     "perl-dap-eval",
+    "perl-dap-command-args",
     "perl-dap-platform",
     "perl-dap-stack",
     "perl-dap-value",
@@ -294,7 +296,7 @@ perl-dap-breakpoint = { path = "crates/perl-dap-breakpoint", version = "0.10.0" 
 perl-dap-eval = { path = "crates/perl-dap-eval", version = "0.10.0" }
 perl-dap-platform = { path = "crates/perl-dap-platform", version = "0.10.0" }
 perl-dap-value = { path = "crates/perl-dap-value", version = "0.10.0" }
-perl-dap-variables = { path = "crates/perl-dap-variables", version = "0.10.0" }
+perl-dap-command-args = { path = "crates/perl-dap-command-args", version = "0.10.0" }
 perl-dap-stack = { path = "crates/perl-dap-stack", version = "0.10.0" }
 perl-feature-catalog = { path = "crates/perl-feature-catalog", version = "0.10.0" }
 perl-lsp-feature-flags = { path = "crates/perl-lsp-feature-flags", version = "0.10.0" }
