+++ crates/perl-dap-command-args/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "perl-dap-command-args"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "Shell-safe command argument formatting for perl-dap"
+readme = "README.md"
+documentation = "https://docs.rs/perl-dap-command-args"
+keywords = ["perl", "dap", "shell", "args", "quoting"]
+categories = ["command-line-interface", "development-tools::debugging"]
+include = ["src/**", "Cargo.toml", "README.md", "LICENSE*"]
+
+[lib]
+name = "perl_dap_command_args"
+path = "src/lib.rs"
+
+[dependencies]
+
+[lints]
+workspace = true
