+++ crates/perl-dap-platform/README.md
@@ -3,4 +3,4 @@
 Cross-platform runtime utilities extracted from `perl-dap`.
 
 This crate provides a focused API for Perl interpreter resolution, path normalization,
-environment setup, and shell-safe argument formatting used by the DAP server.
+and environment setup used by the DAP server.
